Copy static libs into build directory and link them from there
This is an experiment: we copy the libraries from /usr/local/lib into the build directory, and then search for them there when linking. It seems to fix #13, which lends weight to my conjecture that the linker will "helpfully" substitute the dylib versions if it finds them in the path.